+++ src/components/post/postDetail/CommentForm.tsx
@@ -1,5 +1,5 @@
-import { useCommentForm } from '@/hooks/byUse/useCommentForm';
 import { usePostComment } from '@/hooks/queries/byUse/useCommentMutation';
+import { useCommentForm } from '@/hooks/useCustomForm/useCommentForm';
 import { Button } from '@/stories/Button';
 import { UserDetail } from '@/types/postDetailTypes';
 import React, { useEffect } from 'react';
