+++ src/components/post/postDetail/CommentUpdateForm.tsx
@@ -1,5 +1,5 @@
-import { useCommentForm } from '@/hooks/byUse/useCommentForm';
 import { useUpdateComment } from '@/hooks/queries/byUse/useCommentMutation';
+import { useCommentForm } from '@/hooks/useCustomForm/useCommentForm';
 import { Button } from '@/stories/Button';
 import { Comment, UserDetail } from '@/types/postDetailTypes';
 import React, { useEffect, useRef, useState } from 'react';
